data_processing/requirements.txt
# Core
pandas>=1.3.0
numpy>=1.20.0
pymongo>=4.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Google Cloud
google-cloud-storage>=2.0.0
google-cloud-pubsub>=2.8.0
functions-framework>=3.0.0

# Utils
requests>=2.26.0
python-dotenv>=0.19.0
loguru>=0.5.3 